# test: Add direct exception passing for awaits for perf test