Added direct exception passing case for coroutines. Exception are are passed directly after awaited future has been resolved with return_exception_ptr. This option can be a balance between of need frequent exception passing and convenience of using coroutines. The limitations of a such usage the return_exception_ptr cannot be used with void futures.

test                             iterations      median         mad         min         max      allocs       tasks        inst      cycles
chain.then_value                   33390016    29.930ns     0.198ns    29.257ns    30.128ns       1.031       1.062         0.0         0.0
chain.await_value                  21439200    46.201ns     0.055ns    46.146ns    46.681ns       1.063       1.094         0.0         0.0
chain.await_value_as_future        20355264    48.615ns     0.052ns    48.563ns    48.962ns       1.063       1.094         0.0         0.0
chain.then_exception               32634272    30.095ns     0.041ns    30.001ns    30.565ns       1.063       1.062         0.0         0.0
chain.await_exception                538272     1.813us     8.183ns     1.803us     1.834us       2.125       1.094         0.0         0.0
chain.await_exception_as_future    19190272    52.005ns     0.064ns    51.941ns    54.699ns       1.094       1.094         0.0         0.0